requirements.txt
opencv-python
mediapipe
numpy
tensorflow
scikit-learn
Pillow
